requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
Pillow==10.1.0
PyAudio==0.2.11
